Add code to make Terraform run an Ansible playbook against bastioned and non-bastioned hosts when they are started up.
Note that as an example I added both the CyHy bastion host and the
CyHy MongoDB host to the "mongo" group, and the mongo role doesn't yet
do anything useful.  But this shows how it should work, and I wanted
to commit it in case someone else wants to take it further before I
get back to my keyboard.