Append the Terraform workspace only outside of production
Append the Terraform workspace to the MOE bucket and user names only
outside of production.  That way we can share the bucket and users
among production workspaces and the MOE folks won't have to change
anything when we switch from prod-a to prod-b (for example).

I also took the opportunity to add prevent_destroy to the MOE bucket
and the MOE users, and I updated the tags on the MOE bucket.