Add the --upgrade option to the pip command
We need to do this because the apt package python-boto is installed in
our base image, and python-boto in turn pulls in the package
python-requests. That means that when pip runs (via packer) it sees
that requests is already installed. We end up with versions of
requests_aws4auth and requests that do not agree.

Running pip with the --upgrade option forces requests to update to the
latest version even if it is already installed.