Modify the prevent_destroy toggle script for OSX
The -i flag doesn't work the same in the GNU and BSD implementations
of sed, so I have to do some tomfoolery to work around that.  The BSD
implementation of sed also doesn't allow the \s notation I was using
for whitespace matching, so I modified the regex to use [[:space:]]
instead.