Make docker instance use a role not a user for Lambda executions
By using a role I mean giving the EC2 instance an instance profile.
The AWS region where the Lambda functions reside is passed into the
Terraform Ansible so that it can be used to generate an AWS config for
the instance.  (The credentials come from the instance metadata served
locally at http://169.254.169.254, but the configuration still has to
be specified.)