Add NVMe support for the MongoDB instance
This required adding mount options to the scripts/disk_setup.sh
script, so options of "defaults" had to be added to the docker, nmap,
and reporter cloud-init files.  I went ahead and added such an option
to the scripts/non_nvme_disk_setup.sh file as well, so the nessus
cloud-init file had to be updated too.