Improve handling of commander/feeds cyhy.conf conflict
For both the feeds and commander provisioners, if cyhy.conf does not
exist then it is created with the content needed for that particular
service.  If cyhy.conf does exist then the Ansible module blockinfile
is used to insert the appropriate ini section into the file if it is
not already there.