Fix bug in cyhy-feeds provisioning of cyhy.conf file
@KyleEvers pointed out that feeds needs the commander credentials too.
I modified the Ansible code to make sure that they are added, in a way
that will not conflict with what commander's Ansible code does.

I also added blurbs in the feeds and commander Ansible code outlining
the real solution, which is to just create a cyhy.conf with
credentials for each user in mongo_users.other_users.  That will cause
some heartburn, though, since the credential sections will have to be
named using the MongoDB username.  That will necessitate some changes
elsewhere in the code where named sections from cyhy.conf are
explicitly referenced.