Remove reporter instance's explicit dependency on database instance
The reporter doesn't have a service that starts up at boot that will
crash if the database isn't there.  At most there is a cron job that
runs once a week.  Therefore the explicit dependency isn't worth the
trouble because if there are new database and reporter AMIs out there
then it stops us from being able to redeploy the reporter without also
redeploying the database.