Pull the Lambda zips from a bucket instead of the local filesystem
I am making this change because then I can upload the correct zips to
the appropriate bucket.  Then other users need not worry about
generating the zips themselves in another repo, or making sure that
the zips end up in the correct location.

Note that Terraform cannot access buckets that are not in the
provider's region.  This limitation means that we have to create
region-specific buckets.