Allow the user to specify the docker-compose override file for docker and reporter
docker-compose by default reads in the docker-compose.yml and
docker-compose.override.yml files.  By symlinking the proper YML file
to docker-compose.override.yml in the second Ansible, we can install
the proper configuration for the docker and reporter EC2 instances.
In particular, the docker instance will send the BOD 18-01 reports and
the reporter instance will send the Cyber Hygiene reports and the
Cyber Exposure scorecard.