Move redundant else clause
The else clause in a python while loop is called if the loop exits
because the test condition becomes false, and it is skipped if the
loop is exited due to a break statement.  Since the loop has no break
statements, the only way to exit the loop is via the test condition
becoming false.  Therefore the else clause is unneeded here, and its
contents can just follow the while loop.

See here for more details on while loops and else clauses:
https://docs.python.org/3/reference/compound_stmts.html#the-while-statement

This is something that LGTM is warning about, because the else clause
introduces extra complexity that is not needed here.