Move the GitHub OAUTH token from AWS S3 to AWS SSM
This allows us to reference the OAUTH token via the inline aws_ssm
Ansible lookup instead of having to download the file to /tmp from S3,
import its contents into Ansible as a variable, and delete the /tmp
file.

This means that we can get rid of the
cisagov/ansible-role-github-oauth Ansible role and GitHub repo.

See here for more details about the aws_ssm Ansible lookup:
https://docs.ansible.com/ansible/latest/plugins/lookup/aws_ssm.html