Move upgrade and python roles to their own playbooks
Per suggestion this moves the upgrade and python roles into their own
playbooks. This requires updating the Packer configuration for each
instance to reflect that change. This also adds the pip role that has
been in requirements, but not applied to any instances. I also let
our usual pre-commit hooks run against the packer directory, but
ignored ansible-lint results because of the level of effort to
resolve those.