Change temporary pymongo version to 3.6
With Ansible 2.10 I saw issues with the "mongo" provisioner related to pymongo
versions. Per the source:
https://github.com/ansible-collections/community.mongodb/blob/497022063c1feb9416b8632be07c6f2dd8440326/plugins/module_utils/mongodb_common.py#L48-L50
There is a check against the driver and MongoDB version. Since we are using
the 3.6.x version of MongoDB, it will not proceed in the provisioner unless
an appropriate version of pymongo is installed.