Update portscan instance redeploy script
Update the script to use jq to get a list of portscan EC2 instance ids so they
can be terminated all at once. This drastically reduces the runtime of the
script since there is only one Terraaform state search and all desired
terminations are done in on call to AWS.

Added a check for the external programs required for the script to run.