Account for non-existent vulnscan instance in deploy script
Add guardrails to account for a missing instance ID when checking for an
existing ID. A missing instance ID indicates a non-existent instance, so
Terraform items should still be added. This will prevent the script from
bombing out if you were adding new instances or if an instance was down for
some reason (manually stopped on the console, etc).