Add script to update MaxMind GeoIP2 database
This is a utility script I've had laying around that lets you pull the MaxMind
GeoIP2 database locally to update a remote host that does not have internet
access. I've mainly used this to ensure that I'm using the latest GeoIP2 data
before initializing a stakeholder on the server, which is admittedly only done
by the dev team for stakeholders with a large number of hosts.