Remove unnecessary task from vdp_scanner post-Packer
The `orchestrator` job adds `/usr/local/bin` to the PATH for the cron user.
This should no longer be necessary because `docker-compose` is no longer
installed using `pip`, and is instead installed with `apt`. This places the
`docker-compose` script in `/usr/bin` instead, which should be in the cron
user's PATH already.