Remove `local-exec` to terminate EC2 instances
This logic was used to ensure that the EC2 instance for an AWS EBS volume
attachment was in a terminated state so that the volume attachment could be
destroyed successfully. This has not been an issue in more recent COOL
configurations so we are removing it here. In the event that this is still an
problem a relevant issue can be found at:
https://github.com/hashicorp/terraform-provider-aws/issues/4770