Update Nessus instance configuration for TF 0.13
Update the Nessus instance configuration in the terraform/ directory to reflect
changes necessary for Terraform 0.13. This includes removing the dynamic
module configuration and instead leveraging the fact that modules are first-
class citizens in Terraform 0.13. All necessary references for the Ansible
provisioner resources have been updated to reflect this change.