Update mgmt Nessus instance configuration for TF 0.13
Update the management Nessus instance configuration in the terraform/ directory
to reflect changes necessary for Terraform 0.13. This includes removing the
dynamic module configuration and instead leveraging the fact that modules are
first-class citizens in Terraform 0.13. Any references to the previous local
variable are updated to refer to the regular variable that was created as part
of this transition.