Add a variable to control AMI prefixes for Terraform
We have historically changes the prefix for the AMIs created with
Packer when testing so they do not clash with the default of "cyhy".
Using those images has required manually changing the appropriate `.tf`
files. I believe a better solution is to define a variable to specify
different prefixes. This allows a testing configuration to use a mix of
default and testing images for a deployment and leaves it defined in a
centralized location (the corresponding `.tfvars` file).