Fix deploy scripts
Change the database deployment script to target everything at once now
that the Ansible provisioner module can correctly depend on the AWS EC2
instance(s). Change the portscan and vulnscan deployment scripts to fix
the jq queries based on what is in the Terraform state now. Also fix
the resource targeted for the Ansible provisioner to align with changes
made from the transition to Terraform 1.0.