Update cloud-init tasks for all terraform/ instances
As part of setting up instance-specific hostname setting tasks we had
to refactor all the cloud-init configurations in the terraform/
directory. This included eliminating two common tasks and declaring
their equivalent tasks in the configuration for each instance that used
them. With that out of the way we add tasks for each instance to
(courtesy of @jsf9k):
Set the  local hostname to the correct value that will eventually be
obtained from DHCP, since we make liberal use of the
"{local_hostname}" placeholder in our AWS CloudWatch Agent
configuration.

Co-authored-by: Shane Frasier <jeremy.frasier@trio.dhs.gov>