Add the instance log groups to the depends_on clause of the Mongo instances
The CloudWatch Agent running on the Mongo instances will create these
log groups when they start up, but the CloudWatch log metric filters
we create for detecting NVD and KEV sync failures require that they
exist; therefore we create the log groups via Terraform.  At the same
time, we want to avoid a race condition where these instances start up
and create the log groups before the Terraform code can; hence, we add
the log group resources to the depends_on clause for the instances.

Instead of coming up with a clever way to isolate the particular
log group that each instance depends on, I just list the whole
set.  The effect is the same.

Note that this change also necessitated tweaking the
db_instance_hostnames local variable to avoid a cyclic dependency.

Co-authored-by: Nick <50747025+mcdonnnj@users.noreply.github.com>