Change counts to rely on variables instead of instances
Change some `count` arguments to rely on the var.*_instance_count
variables. This allows the EBS volumes to be created independent of the
instances and more importantly allows the instances to be destroyed
while the volumes are being destroyed. This will prevent EBS volume
destruction from getting stuck because the instances aren't being
destroyed.