Set `use_proxy` to false for the Ansible provisioner
The default setting of `use_proxy = true` results in the Ansible
provisioner for Packer creating a local proxy adapter to connect to the
remote instance. This can result in a number of problems with more
recent OpenSSH versions due to changes. We can safely remove the
`ansible.cfg` file we needed for newer OpenSSH versions because the core
problem was the local proxy adapter was set up with an RSA key and with
the proxy disabled the provisioner will instead use the Packer SSH
configuration which is set to use an ED25519 key.