Update the tasks for the cyhy_commander role
Ensure that all filesystem creation tasks have appropriate mode
information. Use a template file instead of a multiline string for a
file's contents. Disable the no-changed-when test for a task since it
will update data if re-run (and there not being an easy way to only run
it under specific circumstances). Update two checks to reflect that the
registered values are booleans already. Consolidate the creation of two
directories into a loop.