Update the tasks for the cyhy_dashboard role
Consolidate directory creation into a loop and remove an unnecessary
path. Ensure the proper mode is applied to created files. Use a
template instead of a multiline string for file contents. Use the copy
module instead of the template module for files that do not use
interpolation. Sort keys for tasks.