Provide the list of nmap and nessus hosts to Ansible
Instead of hard coding the portscan and vulnscan hosts in the commander
configuration we instead generate this list dynamically in the
Terraform configuration and pass it to the cyhy_commander Ansible role.
The Jinja2 template for the commander.conf file has been updated to
support this change. This should allow us to appropriately keep the
commander configuration in sync with the Terraform configuration's
variables.

The list creation is being done in Terraform because doing the same in
the Jinja2 template comes with a number of gotchas and hacky behavior to
achieve the desired output.