Increase the texmf buffer size
TeX uses the buffer to contain input lines, but macro expansion works
by writing material into the buffer and re-parsing the line. As a
consequence, certain constructs require the buffer to be very
large. As distributed by the OS package, the size is 200000, which is
more than sufficient for most documents; however, we have recently
seen a lengthy CyHy report that required a value greater than 500000.