Add a version pin for the community.mongodb collection
The pinned version is the last to officially support MongoDB 3.6 (which
we are still using at this time). If we need changes in versions 1.4.0-
1.5.1 of the collection we will need to update `mongo` Ansible role
defined in this repository to pass `strict_compatibility : false` to any
modules from the community.mongodb collection. However, version 1.5.1
is the last to provide any support for PyMongo <4, so anything beyond
that version would potentially involve significant changes.