Create a base playbook for Packer
This consolidates all of the changes that should be done to all hosts
into a standalone playbook that is the first thing run as part of the
main playbook run against images being built. We install `htop` onto
the bastion AMI because there is no real harm in adding it and it is
part of the standard base configuration established in
cisagov/skeleton-packer.