Update how IMDSv2 is supported for Debian Buster systems
The current method installs python3-botocore from buster-backports.
This can create incompatibilities if another project tries to install
the boto3 package due to a large drift between the versions of the two
packages that are updated in lock-step. This came to light specifically
when trying to install cisagov/cyhy-feeds as a Python 3 package where
the latest version of boto3 was installed as a Python package and it
could not function with the python3-botocore system package's version.