Update S3 configurations in this repository
Add a lifecycle rule to ignore changes to the
server_side_encryption_configuration attribute for all AWS S3 bucket
configurations. This is to support using the back-ported standalone
resource to define that rule for these buckets. Since the
parameter is only deprecated, but not removed, from the aws_s3_bucket
resource we need this rule to ensure consistent applies using the
standalone resource.