Use consistent language in documentation
The scouts have reported that the back-ported holdings are expanding...
I'm not sure how long we'll be able to hold our position against the
creeping tide. Despite the best efforts of @dav3r and @jsf9k, we find
ourselves on the precipice. Only time will tell if we will be able to
hold out against inconsistent language in our projects...

Co-authored-by: dav3r <david.redmin@gwe.cisa.dhs.gov>