Update the Ansible-lint configuration
We need to mock the roles defined in ansible/roles/ as they are not
installed to the locations that Ansible and Ansible-lint look for roles
by default. There is no easy way to modify the locations searched
without overwriting them, so we are mocking them so that Ansible-lint
can pass playbook checks.