Install the system package for pexpect
We prefer to install the system package for Python packages whenever
possible. This is especially important with Debian Bookworm because you
cannot modify the system Python environment directly (with pip) by
default. Switching from installing the pexpect package directly with pip
to installing the python3-pexpect package is a suitable resolution.