Add a new Ansible role to the configuration
This Ansible role writes an AWS CloudWatch Agent configuration file
and restarts the Agent so it takes effect. It offers more configuration
of the log group name than is available with
cisagov/ansible-role-cloudwatch-agent. It is necessary to define it
here instead of enhancing the aforementioned role due to how AMIs are
built and deployed in this configuration. We need to ensure that the
configuration file can be modified for a given Terraform workspace
regardless of whether custom AMIs are used or not.