Add cloud-init to change the CloudWatch Agent configuration file
Add a cloud-init configuration file template that will write a new
configuration file for the CloudWatch Agent. The template will allow
for the configuration of the base of the CloudWatch log group name and
since we have configured the Agent to start after cloud-init has run it
should be updated before any logs have been pushed. Appropriate locals
have been created to provide standardized base log group names for both
the cyhy and bod portions of the configuration. These include the
Terraform workspace being used resulting in the format:
/instance-logs/<Terraform workspace>/(bod|cyhy)