Update the Packer configuration's Ansible requirements
This should ensure that all collections and roles used by the Packer
Ansible configuration are explicitly listed. Note that this required
changing the format of the `requirements.yml` file from a simple list
of roles to a map of collections and roles. This format mirrors the one
already in use for the post-deployment Ansible configuration.