Add a configuration file for terraform-docs
Adds a configuration file for terraform-docs and injection markers to
the README to help ensure consistent population of documentation. This
is not a comprehensive solution because I have found no way to
configure terraform-docs to use the ATX closed header style for
Markdown. It will at least ensure that except for intervention to fix
the header format we have a consistent documentation style.