Change the instance type of the CyHy dashboard
Move from a `c5.xlarge` instance to a `m5zn.xlarge` instance. Based on
observation of the instance running in the production environment I
made the following determinations:
- We are almost maxed on memory so this change takes us from 8GiB to
  16GiB.
- The ncats-webd backend appears to sit on only one vCPU and peg it to
  100% all the time. This change keeps us at four vCPUs but the new
  instance type should have much faster all-core performance.

Additional items of note are that network bandwidth will increase from
"Up to 10Gbps" to "Up to 25Gbps", but EBS bandwidth will decrease from
"Up to 4,750Mbps" to "Up to 3,170Mbps". Since there is not really any
"on-disk" work being done I do not think the EBS bandwidth change will
negatively impact us and the network bandwidth change may improve
performance when talking to the database instance.