Update the security header Lambda's runtime to Node 18
We recently received an email from AWS stating that they are dropping
support for Node 16 in mid-2024.  Now that we are using version 4.9 of
the Terraform AWS provider, we can upgrade to Node 18.  When we move
to the latest version of the Terraform AWS provider, we will be able
to upgrade to at least Node 20.