Add a cron job to restart ncats-webd daily
This adds a configuration to create a cron job in the cyhy_dashboard
Ansible role that will restart the ncats-webd service daily at
~midnight EST. This is done because the team using the dashboard has
been periodically requesting manual restarts of the service due to the
dashboard becoming slow or unresponsive over time. Once the service is
restarted the dashboard returns to normal usability. Hopefully
restarting it daily will be a tolerable band-aid for the performance
issues.