Fix DHCP4 configuration in Netplan
Debian Bookworm switches DNS management from resolvconf to
systemd-resolved. This change results in Bookworm instances having an
incompatible hostname resolution configuration. Since we rely on
cloud-init to automatically configure some of our DNS settings we need
to adjust the configuration of Netplan (used by cloud-init) to get the
correct configuration for our system.

The issue is that Netplan uses a default of false for the value of
dhcp4-overrides.use-domains and cloud-init does not explicitly set this
key or provide a means to do so. We remedy this by modifying the
cloud-init configuration of Bookworm instances to use a Python script
to adjust the Netplan configuration and then re-apply Netplan to enable
our desired configuration.

Co-authored-by: Shane Frasier <jeremy.frasier@gwe.cisa.dhs.gov>