Remove client certificate results sending
This removes the use of the cisagov/client-cert-update Docker image in
the CyHy environment. That includes removing the use of
cisagov/ansible-role-client-cert-update, the client_cert_update role in
this project, and any scaffolding to support use of these two Ansible
roles. This is done because the email sent by this Docker configuration
does not appear to be in need any longer and so it makes sense to
remove the functionality from this project to improve maintainability.