Install the geoipupdate tool on the database AMI
This adds installation of the MaxMind geoipupdate tool as well as
configuring automatic updates on the database AMI. This will ensure
that the MaxMind database is kept up-to-date. This is the only AMI
that needs this configured since, when deployed, it sometimes has
commands run on it that use the MaxMind data. The other deployed
AMIs are not in similar situations and only have it installed as a
dependency of needing the cisagov/cyhy-core Python package.