Stop forcing Python 2 on remotes in our Packer configurations
Currently we force Ansible to discover Python 2 and use it (if present)
in the dashboard, mongo, and reporter AMI configurations. This now
proves problematic on ansible-core 2.16 versions 2.16.5+ due to a
backported change that breaks Python 2 support. It also makes it
impossible for us to use ansible-core version 2.17.

We have already ensured that all of our Ansible roles declare
the Python 2 version of pip if needed. This should allow us to use the
Python 3 interpreter auto-discovered by Ansible to run against remotes
when building AMIs for _all_ image configurations.