Add an arm64 bastion source
Add a source configuration to build an arm64 versions of the bastion
AMI. In addition the x86_64 and build configurations are broken out
into their own files and the build configuration is updated to include
the arm64 source. The Packer template README is updated to mention the
change in building a specific AMI if it has multiple architectures
available.