Use a GitHub token for Packer when running pre-commit
This pulls in the changes made in cisagov/skeleton-packer#437. They
apply here when pre-commit is run because the `packer_validate` hook
will attempt to run a `packer init`. I have seen intermittent failures
of this hook without any output to indicate that there are problems
with the Packer template. I believe we may be running into the same
issue around unauthenticated requests during the `packer init` run.

Co-authored-by: Jeremy Frasier <jeremy.frasier@gwe.cisa.dhs.gov>