Fix line lengths in the cyhy_commander Ansible role
Convert single line strings that violate the new yamllint line length
rules into multiline strings. We must also disable the line-length
rules for a task because the offending line is part of a block
statement and there is no good way to reformat or disable the rule for
just that line.