Fix the Nessus activation code `regex_search`
In #856 the line that extracts the Nessus activation code was turned
into a multiline string. However, testing for that pull request did not
include running the provisioner against an already deployed instance.
When testing other changes I did so and found that the `regex_search`
did not function as expected. This resolves that by adjusting how the
argument is escaped.