Add a second Docker subnet
This must be done because our serverless ElastiCache Valkey cache
requires 2-3 subnets in different availability zones.  We use the
cisagov/distributed-subnets-tf-module Terraform module to satisfy this
latter requirement.

Note also that any EBS volumes that will be attached to EC2 instances
in the Docker subnets must exist in the same availability zone as the
instance.  Hence, when deploying this code to an already-existing
deployment it may be necessary to recreate existing data volumes by:
1. Manually snapshotting existing EBS volumes
2. Manually creating a new EBS volume in the desired availability zone
from the snapshot
3. Importing the new data volume into the Terraform state
4. Applying the Terraform code
5. Manually deleting the original data volume