Support Graviton `portscan` instances
Add the ability to specify the number of ARM64 and AMD64 `portscan`
instances to deploy. The current configuration will deploy AMD64-based
instances first and then ARM64-based instances. This was done to make
the implementation as easy as possible given all of the other elements
of the configuration that interact with these instances.