Pull files from a specific version of cisagov/cyhy-core
Instead of pulling the places helper files from the `develop` branch,
we add a new Ansible variable to specify the version to pull down with a
default value of the latest version tag. A corresponding Terraform
variable is added to pass through a value to this new, required Ansible
variable.